requirements.txt
bcrypt==3.1.1
certifi==2016.9.26
cffi==1.8.3
click==6.6
Flask==0.11.1
Flask-Bcrypt==0.7.1
Flask-Login==0.3.2
Flask-SQLAlchemy==2.1
future==0.15.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.11.1
pandas==0.18.1
pycparser==2.14
python-dateutil==2.5.3
python-telegram-bot==5.1.0
pytz==2016.6.1
requests==2.11.1
scikit-learn==0.18
scipy==0.18.1
six==1.10.0
sklearn==0.0
slugify==0.0.1
spotipy==2.3.8
SQLAlchemy==1.0.15
urllib3==1.18
Werkzeug==0.11.11
